basic greedy assignment combination
