+++ pil_analyzer/src/lib.rs
@@ -9,6 +9,7 @@ use std::{collections::HashMap, fmt::Display};
 
 use number::{DegreeType, FieldElement};
 pub use parser::ast::{BinaryOperator, UnaryOperator};
+
 use util::expr_any;
 
 pub fn analyze<T: FieldElement>(path: &Path) -> Analyzed<T> {
